small code cleanup
